third_party/blink/web_tests/editing/inserting/text-into-anchor-in-table.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    [
      '<div contenteditable><table><tbody><tr><td>',
        '<a href="http://google.com/">',
          '<b>foo|bar</b>',
        '</a>',
      '</td></tr></tbody></table></div>'
    ].join(''),
    'insertHTML baz',
    [
      '<div contenteditable><table><tbody><tr><td>',
        '<a href="http://google.com/">',
          '<b>foo</b>',
          'baz|<b>bar</b>',
        '</a>',
      '</td></tr></tbody></table></div>'
    ].join(''));
}, 'insertHTML text into anchor in table');
</script>
